requirements.txt
Django==4.1.1
Pillow==9.2.0
django-tables2==2.4.1
django-crispy-forms==1.14.0
gunicorn==20.1.0
python-dotenv==0.21.0
pyyaml
jinja2